refactor: move RunConfig out of cli/
agent/loader.py imported from cli.run_config, a forward-breach
where agent core depended on CLI. Move the module to the package
root so non-CLI consumers (agent, services, dashboard) no longer
reach into cli/ for a config primitive.

All seven import sites and three test files updated in this commit;
no shim left behind.
